Catalog - batch operations.

+++ web/shk-app/src/app/page-table.abstract.ts
@@ -145,7 +145,7 @@ export abstract class PageTableAbstractComponent implements OnInit {
     title: string = 'Page with data table';
     modalRef: NgbModalRef;
     loading: boolean = false;
-    selectedIds: string[] = [];
+    selectedIds: number[] = [];
     collectionSize = 0;
     queryOptions: QueryOptions = new QueryOptions('name', 'asc', 1, 10, 0, 0);
 
@@ -195,28 +195,72 @@ export abstract class PageTableAbstractComponent implements OnInit {
             if( result == 'accept' ){
                 this.deleteItem(itemId);
             }
-        }, (reason) => {
-
         });
     }
 
-    deleteItem(itemId: number): void{
-        this.dataService.deleteItem(itemId)
-            .then((res) => {
-                if(res.success){
-                    this.getList();
-                } else {
-                    if(res.msg){
-                        this.modalRef = this.modalService.open(AlertModalContent);
-                        this.modalRef.componentInstance.modalContent = res.msg;
-                        this.modalRef.componentInstance.modalTitle = 'Error';
-                        this.modalRef.componentInstance.messageType = 'error';
-                    }
+    confirmAction(message: string) {
+        this.modalRef = this.modalService.open(ConfirmModalContent);
+        this.modalRef.componentInstance.modalTitle = 'Confirm';
+        this.modalRef.componentInstance.modalContent = message;
+        return this.modalRef.result;
+    }
+
+    blockSelected() {
+        if(this.selectedIds.length === 0){
+            this.showAlert('Nothing is selected.');
+            return;
+        }
+        console.log('blockSelected', this.selectedIds);
+    }
+
+    deleteSelected() {
+        if(this.selectedIds.length === 0){
+            this.showAlert('Nothing is selected.');
+            return;
+        }
+        this.confirmAction('Are you sure you want to delete all selected items?')
+            .then((result) => {
+                if( result == 'accept' ){
+                    this.dataService.deleteByArray(this.selectedIds)
+                        .then((res) => {
+                            if(res.success){
+                                this.getList();
+                            } else {
+                                if(res.msg){
+                                    this.showAlert(res.msg);
+                                }
+                            }
+                        });
                 }
             });
     }
 
-    actionRequest(actionValue : [string, number]): void {
+    showAlert(message: string) {
+        this.modalRef = this.modalService.open(AlertModalContent);
+        this.modalRef.componentInstance.modalContent = message;
+        this.modalRef.componentInstance.modalTitle = 'Error';
+        this.modalRef.componentInstance.messageType = 'error';
+    }
+
+    deleteItem(itemId: number): void{
+        this.confirmAction('Are you sure you want to remove this item?')
+            .then((result) => {
+                if( result == 'accept' ){
+                    this.dataService.deleteItem(itemId)
+                        .then((res) => {
+                            if(res.success){
+                                this.getList();
+                            } else {
+                                if(res.msg){
+                                    this.showAlert(res.msg);
+                                }
+                            }
+                        });
+                }
+            });
+    }
+
+    actionRequest(actionValue: [string, number]): void {
         switch(actionValue[0]){
             case 'edit':
                 this.modalOpen(actionValue[1]);
@@ -225,7 +269,7 @@ export abstract class PageTableAbstractComponent implements OnInit {
                 this.modalOpen(actionValue[1], true);
                 break;
             case 'delete':
-                this.deleteItemConfirm(actionValue[1]);
+                this.deleteItem(actionValue[1]);
                 break;
             case 'changeQuery':
                 this.getList();
